Update ARM platform data types in int.h

+++ src/platforms/arm/int.h
@@ -1,5 +1,8 @@
 #pragma once
 
+#include <stdint.h>
+#include <stddef.h>
+
 namespace fl {
     // ARM platforms (32-bit): short is 16-bit, long is 32-bit
     // uint32_t resolves to 'unsigned long' on most ARM toolchains
@@ -20,7 +23,7 @@ namespace fl {
     typedef long long i64;
     typedef unsigned long long u64;
     // size_t is unsigned long on ARM (32-bit)
-    typedef unsigned long sz;
+    typedef size_t sz;
     // uintptr_t is unsigned long on ARM (32-bit pointers)
-    typedef unsigned long uptr;
+    typedef uintptr_t uptr;
 }
